+++ RUFAS/output_manager.py
@@ -332,7 +332,47 @@ class OutputManager(object):
         else:
             pool[key]["values"].append(deepcopy(value))
 
-    def add_variable(self, name: str, value: Any, info_map: dict[str, Any], first_info_map_only: bool = False) -> None:
+    def _map_simulation_day(
+        self, info_map: dict[str, Any], overwrite: bool = False, simday: int | None = None
+    ) -> dict[str, Any]:
+        """
+        Create an update info map that include simulation_day
+
+        Parameters
+        ----------
+        info_map: dict[str, Any]
+            original infomap to check and/or update the simulation_day element
+        overwrite: bool, default False
+            should simulation_day be overwritten if it already exists in info_map?
+        simday: optional int
+            the simulation day to add (or overwrite) to the infomap
+
+        Returns
+        -------
+        the info map, updated to include simulation_day, if it is available.
+        """
+        imap_copy = dict(info_map)
+
+        time_day = self.time.simulation_day if hasattr(self.time, "simulation_day") else None
+
+        day_to_use = simday if simday is not None else time_day
+
+        should_add_sim_day = overwrite or "simulation_day" not in imap_copy.keys()
+
+        if day_to_use is not None and should_add_sim_day:
+            imap_copy["simulation_day"] = day_to_use
+
+        return imap_copy
+
+    def add_variable(
+        self,
+        name: str,
+        value: Any,
+        info_map: dict[str, Any],
+        first_info_map_only: bool = False,
+        overwrite_simulation_day: bool = False,
+        simulation_day: int | None = None,
+    ) -> None:
         """
         Adds a variable to the pool.
 
@@ -358,6 +398,14 @@ class OutputManager(object):
         first_info_map_only : bool, default False
             If true, records only the first info map passed for that variable. If false, records all info maps passed
             for that variable.
+        overwrite_simulation_day: bool, default False
+            if true, the variable's info map is automatically updated with the current simulation_day or
+            (self.time.simulation_day if 'simulation_day = None'), otherwise the simulation_day specified
+            within the info_map is used if available.
+        simulation_day: optional int
+            if present, this simulation_day will be used, otherwise the simulation_day will be taken from the time
+            attribute, if present. This option is provided to allow variables created outside the main simulation
+            loop (i.e., herd initialization) to be padded.
 
         Raises
         ------
@@ -374,8 +422,10 @@ class OutputManager(object):
             raise KeyError(f"'units' missing in units dict for a variable in {name}.")
         units = self._stringify_units(units)
 
-        key = self._generate_key(name, info_map)
-        self._add_to_pool(self.variables_pool, key, value, {**info_map, "units": units}, first_info_map_only)
+        updated_imap = self._map_simulation_day(info_map, overwrite_simulation_day, simulation_day)
+
+        key = self._generate_key(name, updated_imap)
+        self._add_to_pool(self.variables_pool, key, value, {**updated_imap, "units": units}, first_info_map_only)
 
         if isinstance(value, dict):
             for k, v in value.items():
@@ -394,7 +444,10 @@ class OutputManager(object):
                 self._save_current_variable_pool()
 
     def add_variable_bulk(
-        self, variables: list[tuple[dict[str, Any], dict[str, Any]]], first_info_map_only: bool = False
+        self,
+        variables: list[tuple[dict[str, Any], dict[str, Any]]],
+        first_info_map_only: bool = False,
+        overwrite_simulation_day: bool = False,
     ) -> None:
         """
         Iterate through all variables and call add_variable() on each of them.
@@ -406,10 +459,12 @@ class OutputManager(object):
             being the variable name and the value being the output value, and its corresponding info map.
         first_info_map_only : bool, default False
             If true, records only the first info_map passed for each variable.
+        overwrite_simulation_day: bool, default False
+            Passed to add_variable(). If true, any simulation_day value provided in the info_maps is overwritten.
         """
         for variable, info_map in variables:
             name, value = list(variable.items())[0]
-            self.add_variable(name, value, info_map, first_info_map_only)
+            self.add_variable(name, value, info_map, first_info_map_only, overwrite_simulation_day)
 
     def _save_current_variable_pool(self) -> None:
         """